Auto-start support for API23 devices
On API23 there's no Always-On VPN mode, so we do things a bit differently.
Instead of relying on the OS to start our vpn service, we do it ourselves
while handling the BOOT_COMPLETED intent.

This isn't something that's really feasible to do on newer devices (what can happen
in background services is pretty tightly restricted), but this approach works well
on older devices.

Actual "startup flow" is entirely shared with the Always-On startup logic.

End-result is that we launch the VPN service as it was last-configured by the user
shortly after booting.